Converted some of the memcached API's to use size_t instead of unsigned int.

+++ libmemcached/common.h
@@ -136,7 +136,7 @@ void server_list_free(memcached_st *ptr, memcached_server_st *servers);
 
 LIBMEMCACHED_LOCAL
 memcached_return memcached_key_test(const char **keys, size_t *key_length,
-                                    unsigned int number_of_keys);
+                                    size_t number_of_keys);
 
 
 LIBMEMCACHED_LOCAL
